Rename errors directory to error-handlers
