+++ kiosk/src/js/application_updater.js
@@ -5,7 +5,7 @@ const cron = require('node-cron');
 const NotificationListener = require('./helpers/notification_listener');
 const PropertiesLoader = require('./helpers/properties_load_helper');
 const UserInteractionsManager = require('./user_interactions_manager');
-const NetworkErrorsHandlingService = require('./services/errors/network_errors_handling_service');
+const NetworkErrorsHandlingService = require('./services/error-handlers/network_errors_handling_service');
 const AutoupdateScheduleWatcher = require('./services/data/autoupdate_schedule_watcher');
 const ConnectionStatusService = require('./services/network/connection_status_service');
 const Logger = require('./services/logger/logger');
