+++ kiosk/src/js/services/error-handlers/global_errors_handling_service.js
