+++ kiosk/src/js/services/error-handlers/network_errors_handling_service.js
