+++ kiosk/src/js/services/error-handlers/uncaught_errors_handling_service.js
