+++ kiosk/src/js/services/services_initialiser.js
@@ -4,9 +4,9 @@ class ServicesInitialiser {
     //should be called ON application initializing
     static initBaseServices() {
         require('./logger/logger').setupLoggerProperties();
-        require('./errors/global_errors_handling_service');
-        require('./errors/network_errors_handling_service');
-        require('./errors/uncaught_errors_handling_service');
+        require('./error-handlers/global_errors_handling_service');
+        require('./error-handlers/network_errors_handling_service');
+        require('./error-handlers/uncaught_errors_handling_service');
         require('./network/connection_status_service');
         require('./data/autoupdate_schedule_watcher');
         require('./data/server_data_watcher');
